Focus-timer.Chat-Bot-master/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
gevent==23.9.1
gevent-websocket==0.10.1
google-generativeai==0.3.2
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
itsdangerous==2.1.2
gunicorn==21.2.0
